+++ src/main/java/com/healthbest/api/bloodSugar/BloodSugar.java
@@ -1,4 +1,4 @@
-package com.healthbest.api.domain;
+package com.healthbest.api.bloodSugar;
 
 import com.healthbest.api.user.domain.User;
 import jakarta.persistence.*;
