+++ src/main/java/com/healthbest/api/user/domain/vo/Authority.java
@@ -1,0 +1,6 @@
+package com.healthbest.api.user.domain.vo;
+
+public enum Authority {
+    ROLE_USER,
+    ROLE_ADMIN
+}
