+++ src/main/java/com/healthbest/api/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.healthbest.api.user.repository;
+
+import com.healthbest.api.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.lang.reflect.Member;
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByLoginId(String loginId);
+    boolean existsByLoginId(String loginId);
+}
